refactor(startup): consolidate src/init into src/boot
Boot logic lived in three places — src/init.tsx, src/init/, and the new
src/boot/ — which is harder to follow than what it replaced. Everything the
boot path owns now lives under src/boot/:

  src/init.tsx              -> src/boot/bootstrap.tsx
  src/init/keycloak.ts      -> src/boot/keycloak.ts
  src/init/tabManager.ts    -> src/boot/tabManager.ts
  src/init/googleAnalytics  -> src/boot/googleAnalytics.ts
  src/init/AppBootstrap.tsx -> src/boot/AppBootstrap.tsx
  src/init/init_docs/       -> src/boot/boot_docs/

Pure move: no logic changes. Importers of keycloak (11 of them, all feature
modules reaching for KeycloakContext) now use the @/boot/keycloak alias.
vite.config.ts's init-chunk modulepreload match and src/index.tsx's dynamic
import follow the new path — both verified against a real production build,
since the plugin's match is a string comparison that fails silently.

204 test files / 2659 tests pass; build verified to still inject the boot
shell as the first module script.
