+++ src/boot/tabManager.ts
