+++ src/features/Login/LoginButton.tsx
@@ -6,7 +6,7 @@ import { ReactElement, useContext, useState } from 'react'
 import { AppConfigContext } from '../../AppConfigContext'
 import { useCredentialStore } from '../../data/hooks/stores/CredentialStore'
 import { logApi } from '../../debug'
-import { KeycloakContext } from '../../init/keycloak'
+import { KeycloakContext } from '@/boot/keycloak'
 import { LoginPanel } from './LoginPanel'
 
 export const LoginButton = (): ReactElement => {
