+++ src/features/SummaryPanel/NetworkPropertyPanel.tsx
@@ -18,7 +18,7 @@ import { useMessageStore } from '../../data/hooks/stores/MessageStore'
 import { useViewModelStore } from '../../data/hooks/stores/ViewModelStore'
 import { useWorkspaceStore } from '../../data/hooks/stores/WorkspaceStore'
 import { useSaveCurrentNetworkToNDEx } from '../../data/hooks/useSaveCurrentNetworkToNDEx'
-import { KeycloakContext } from '../../init/keycloak'
+import { KeycloakContext } from '@/boot/keycloak'
 import { IdType } from '../../models/IdType'
 import { MessageSeverity } from '../../models/MessageModel'
 import { NetworkSummary } from '../../models/NetworkSummaryModel'
