+++ src/features/ToolBar/DataMenu/CopyNetworkToNDExMenuItem.tsx
@@ -19,7 +19,7 @@ import { useWorkspaceStore } from '../../../data/hooks/stores/WorkspaceStore'
 import { useLoadNetworkSummaries } from '../../../data/hooks/useLoadNetworkSummaries'
 import { useSaveCyNetworkCopyToNDEx } from '../../../data/hooks/useSaveCyNetworkCopyToNDEx'
 import { logUi } from '../../../debug'
-import { KeycloakContext } from '../../../init/keycloak'
+import { KeycloakContext } from '@/boot/keycloak'
 import { NetworkSummary } from '../../../models'
 import { IdType } from '../../../models/IdType'
 import { MessageSeverity } from '../../../models/MessageModel'
