+++ src/features/ToolBar/DataMenu/LoadFromNdexDialog.tsx
@@ -48,7 +48,7 @@ import { useMessageStore } from '../../../data/hooks/stores/MessageStore'
 import { useNetworkSummaryStore } from '../../../data/hooks/stores/NetworkSummaryStore'
 import { useWorkspaceStore } from '../../../data/hooks/stores/WorkspaceStore'
 import { logUi } from '../../../debug'
-import { KeycloakContext } from '../../../init/keycloak'
+import { KeycloakContext } from '@/boot/keycloak'
 import { IdType } from '../../../models/IdType'
 import { MessageSeverity } from '../../../models/MessageModel'
 import { NetworkSummary } from '../../../models/NetworkSummaryModel'
