+++ src/features/ToolBar/DataMenu/SaveWorkspaceToNDEx.tsx
@@ -3,7 +3,7 @@ import React, { useContext, useState } from 'react'
 import { AppConfigContext } from '../../../AppConfigContext'
 import { useCredentialStore } from '../../../data/hooks/stores/CredentialStore'
 import { useWorkspaceStore } from '../../../data/hooks/stores/WorkspaceStore'
-import { KeycloakContext } from '../../../init/keycloak'
+import { KeycloakContext } from '@/boot/keycloak'
 import { BaseMenuItemProps } from '../BaseMenuItemProps'
 import { DropdownMenuItem } from '../DropdownMenu'
 import { WorkspaceNamingDialog } from './WorkspaceNamingDialog'
